emotion_recognition/requirements.txt
Pillow
torch
torchvision
flask
flask-cors